chore: split styles
